requirements.txt
pyowm==2.10.0
python-dotenv==0.12.0